pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pdf-extractor"
version = "0.1.0"
description = "A Python tool for extracting text and tabular data from PDF files"
authors = [
    {name = "Developer", email = "developer@example.com"},
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "polars>=0.20.0",
    "PyPDF2>=3.0.0",
    "pdfplumber>=0.10.0",
    "tabula-py>=2.8.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "reportlab>=4.0.0",
    "pyarrow>=17.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/example/pdf-extractor"
Repository = "https://github.com/example/pdf-extractor"

[project.scripts]
pdf-extractor = "pdf_extractor.cli:main"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.ruff]
target-version = "py38"
line-length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
